% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem-single.r
\name{mem_single}
\alias{mem_single}
\title{Fit the MEM Model for single subgroup}
\usage{
mem_single(responses, size, name, drug_index, p0 = 0.15, shape1 = 0.5,
  shape2 = 0.5, prior = rep(1, length(responses))/2,
  hpd_alpha = 0.05, alternative = "greater", call = NULL)
}
\arguments{
\item{responses}{the number of responses in each basket.}

\item{size}{the size of each basket.}

\item{name}{the name of each basket.}

\item{drug_index}{the index of the basket to be studied.}

\item{p0}{the null response rate for the poster probability calculation
(default 0.15).}

\item{shape1}{the first shape parameter(s) for the prior of each basket
(default 0.5).}

\item{shape2}{the second shape parameter(s) for the prior of each basket
(default 0.5).}

\item{prior}{the matrix giving the prior inclusion probability
for each pair of baskets. The default is on on the main diagonal and 0.5
elsewhere.}

\item{hpd_alpha}{the highest posterior density trial significance.}

\item{alternative}{the alternative case definition (default greater)}

\item{call}{the call of the function (default NULL).}
}
\description{
Fit the MEM model for single subgroup using full Bayesian inference.
}
\examples{
# 6 baskets, each with enrollement size 5
trial_sizes <- rep(25, 6)

# The response rates for the baskets.
resp_rate <- 0.15

# The trials: a column of the number of responses and a column of the
# the size of each trial.
trials <- data.frame(
  responses = rbinom(trial_sizes, trial_sizes, resp_rate),
  size = trial_sizes,
  name = letters[1:6]
)

mem_single(trials$responses, trials$size, trials$name, 2)
}
