% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_borrow_density}
\alias{plot_borrow_density}
\title{Plot the posterior density of the fitting result}
\usage{
plot_borrow_density(x, ...)
}
\arguments{
\item{x}{the fitted model calculated from the borrow_single() and borrow_multiple() function.}

\item{...}{other parameters}
}
\description{
Create the posterior density plot of the studied basket.
}
